Reduce HTTP timeout for mailcow API to 5 seconds
requests timeout is the time a server is allowed not to respond at all,
i.e. we only abort the connection if there is no byte received
for this duration. For this purpose 30 seconds is too much,
it is better to timeout after 5 seconds and restart from scratch.
Otherwise within 60 seconds allowed for processing a single
new_email requests we can only do 2 attempts, while the limit
is 10 attempts.